backend/requirements.txt
fastapi[all]
requests
minio
cython
celery
python-multipart 
pymongo
redis
python-logging-loki
uvicorn
ultralytics
websockets